building: Ensure DEPENDENCY's correctly get MERGE'd
The common_prefix has already been normcase'd, so ensure that
the path we are trying to remove the common_prefix from is
also normcase'd so that replace can work.

When defining name of the dependency get the basename from
tpl[1] rather than using tpl[0] as tpl[0] may have not have
the proper casing (it may have been normcase'd) which will
cause issues on operating systems where filename are
case-sensitive (eg Linux).