Hooks: Make the soundfile hook platform aware
The hook exists to package the libsndfile binary when required.
However pysoundfile only bundles libsndfile on OSX and Windows,
on Linux it relies on a system package. Attempting to include
_soundfile_data on Linux therefore results in an unexpected
exception as per #4325.

Furthermore the bundled files are different on OSX and Windows
so the data required has to be platform specific.