Bootloader build: wscript: Make arch compilers flags optional ... if the arch (bit-size) is auto-detected.
The arch-flags are still required to be supported if option `--target-arch` is
given.

This assumes: If the auto-detection (which runs without arch-flags) reports
some bit-size, the compiler is expected to generate code for this bit-size
without requiring any arch-flags.

This should solve a series of issues related to compilers not supporting
arch-flags.

See #4278 or 2d5898c41efed11272790e96f478451ea36d64d6.