Tests, building, depend: Rework caching of PyiModuleGraph.
Caching of PyiModuleGraph is only relevant for running the test-suite
or when running PyInstaller programmatically.

Since the last commit moved searching the normal (post-graph) hooks
into PyiModulegraph, additional hook dirs were ignored in the
test-suite and respective tests failed.

This showed that the implementation of that cached graph was not
clean:
- fixture pyi_modgraph created the graph (using initialize_modgraph)
- AppBuilder._test_building() added it to CONF
- Analysis.assemble() checked whether to cached graph could be used.

Now initialize_modgraph() returns the cached graph if applicable. See
comment there for details.

As a side-effect we get rid of deep-copying the graph which caused
curious errors in ModuleHookCache related to weakref showing up
after the last commit.