hooks: Fix distutils detection if virtualenv is used.
Don't rely on virtualenv internals for system distutils detection, fixes #4064 

When run in a venv, `distutils` present in the venv is not suitable
for freezing, instead the system-wide `distutils` should be used.
`pyinstaller` relies on a pre-find-module-path hook to change the
search path to point to this system-wide `distutils` location.

To do this, it made use of a `distutils_path` attribute that was set
in `virtualenv`, which pointed to the system-wide `distutils` dir.
This `distutils_path` was never meant to be part `virtualenv`'s public
api. Consequently, as part of a change in virtualenv release 16.3
(https://github.com/pypa/virtualenv/pull/1289), this `distutils_path`
points to the `__init.py__` of the system-wide `distutils` when run on
python >= 3.4, as opposed to the directory. Consequently, the
`os.path.dirname` used in the original hook returns the wrong result.

Rather than continuing to rely on the `distutils_path` attribute, the
hook now figures out where the system-wide python libs live
using, the same technique `virtualenv` uses (by checking for the
`opcode` module location).