Depend: Handle explicit namespace packages correctly
Explicit namespace packages contain a `__init__.py` file and use either
`pkgutil.extend_path` or `pkg_resources.declare_namespace`. These are
currently not handled by PyInstaller.

A namespace package is only special in the sense that it has multiple
entries in the `__path__` attribute, which means that without actually
importing the package the only way to tell if it is a namespace package
is by looking at the contents of the `__init__.py` file. This is
implemented based on the code in
https://github.com/PyCQA/astroid/blob/6076874993b26b141f4adfbcf2a48e703d84298f/astroid/interpreter/_import/spec.py#L236.