Tests/CI: Avoid binary operator rule conflict with flake8
As flake8 now enables both the W503 and W504 rules it is not possible to
break an expression containing binary operators on multiple lines. As
one of the rules needs to be white listed I've added W504 to the ignore
list as W504 is the recommendation for the future.

See https://lintlyci.github.io/Flake8Rules/rules/W504.html.