setup: Add install extras for explicitly installing platform-specific dependencies
Commit 8365f1b of pull-request Pandinosaurus/pyinstaller#8 made
platform-specific package dependencies conditional within the setuptools script
in order to avoid installing unnecessary platform-specific dependencies.

While this default setup script behavior is desirable it can be problematic in
projects that depend on PyInstaller and pin package dependency versions. If
such a project wishes to target multiple platform compile targets it will fail
because of missing conditional platform-specific dependencies. Adding install
extras is a convenient way to remedy for this problem for projects that will
knowingly need to support more than a single platform compile target.