Test/CI: Enable default_hook tests in test suite
Add the `--use-default-hook` flag as an additional parameter to the
parametrized functional tests in `test_import.py`. This will cause each
`onedir` and `onefile` test to run one more time with the
`--use-default-hook` parameter as well.

We have to skip the `test_nspkg3_bbb_zzz` test due to a bug in ZipImport
which causes the `import nspkg3.bbb` statement to fail. This can be
reproduced by adding the contents of the nspkg3-pkg directory to
`sys.path` and attempting to import any of the eggs: `nspkg3.bbb` will
fail if there are any other sub-packages from that namespace on the
Python path as well. This fails the test in `default_hook` mode as we
call `collect_all` during the packaging process and to determine the
`__file__` attribute of the egg we have to import it.