building: Ignore ImportError in is_package
The `is_package()` function looks up the loader and then calls
`is_package()` on the loader itself to determine if the package in
question is a module or a package. We currently only consider the
failure case of finding a loader for the package, but it can also happen
that we do find a loader, but the loader won't be able to import the
module when we call `is_package()` on it. This happens with
`importlib._bootstrap` which cannot be directly imported.

With this change we handle such import failures by returning False,
which is consistent with the existing failure mode when we cannot find a
loader.