Test/CI: Run all onedir tests with default hook
Add a new Travis job to run all PyInstaller tests with the
`--use-default-hook` command line argument. This is done in a separate
job to avoid running over the Travis per-job time limit and is limited
to the `onedir` fixture type as running in `onedir` vs `onefile` should
not affect this feature.