building: Allow usage of VSVersionInfo again.
fixes #4381: The option to pass a VSVersionInfo object as version
argument to EXE was a feature that was explicitly implemented,
but was overlooked when attempts were made to make paths relative.
The bug is fixed by checking whether the object passed as version
argument is a VSVersionInfo object. This change explicitly only
checks if it is a VSVersionInfo object, if not it is still assumed
to be a path-like object.