Bootloader build: Fix .rc.o file not found with MinGW-w64.
Fixes issue #4501. When trying to build the bootloader with MinGW-w64,
it was giving a windres.exe: windows\run.rc.o: No such file or directory
error. This is because waf only sets the target and source arguments,
WINRC_TGT_F and WINRC_SRC_F, when there is no WINRC configuration
already set. PyInstaller with MinGW-w64 is setting the WINRCFLAGS which
changes the WINRC configuration. The fix for this is to set the
WINRC_TGT_F and WINR_SRC_F values manually.

Signed-off-by: Dan Yeaw <dyeaw@ford.com>