Bootloader: Add generic API for splash screen support.
Add a new skeleton feature into the bootloader.
A generic API for an platform-independent usage
is defined, which should call - based on the
target distribution - the methods for creating
a splash window. The bootloader and the Python
interpreter should communicate with each other
via an IPC mechanism. The platform dependent
implementations will be added in later commits,
this is just the structure.