Bootloader: Add support for an splash screen on Windows.
Implements a iOS style splash screen for Windows into the
generic splash screen api. The implementation expects a SPLASH
structure within the appended CArchive. It is capable of
displaying either a 24bit or 32bit bitmap. At 32bit
transparent pixels are supported, which means splash screens
of every shape can be displayed.
Required libraries (GDI32) for window managing in Windows
are added to the build system.