Hooks: Add the PyInstaller module "pyi_splash".
Add the pyi_splash module that connects to the IPC server of the
bootloader and acts as an RPC interface for it. This module can
send commands to the splash screen, enforcing a communication
protocol. The module has the advantage that the PyInstaller user is
independent of the communication method between the Python
interpreter and the bootloader. Since it is an "internal" PyInstaller
module and therefore cannot be installed by a package manager,
modulgraph finds it by using two hooks.