building: Add a building target and writer for Splash.
Add a building target for the splash resources, which can
be called from a spec file. This module creates the necessary
resource file for the splash screen feature of the bootloader.
The build target has its own resource writer that is not based
on ArchiveWriter, as it is not an archive, but a binary file.
The splash build target can be added as an argument to the
build target EXE like a TOC or Target. It also adds a command-
line option '--splash IMAGE_FILE', which inserts a splash target
into the spec file.