Bootloader build: wscript: Properly check for 32-/64-bit compiler flags.
* First only check the compiler-flag, without linking.
* Then separately check the linker-flag.

When using only one check (for the compiler-flags), I was hit by the
arch-flag not been passed to the linker. It took me quite some time to
figure out that one either needs to pass `linkflags`, too, or needs to
compile only (and another quite some time to figure out how to achieve
this).

I then decided to use two separate tests for compilee- and link-flags,
since if passing both, waf reports "Checking for *linker* flags". Which
is a clear hint on "please check the compiler-flags first".