Bootloader: pyi_path_join no longer allocates result buffer.
Allocating the result buffer would be a good thing, since this would
circumvent path length. limitations. But if we use it, we should change all of
the bootloader to do so, and this had shown to be quite a lot, lot of work.

So, since we do allocate the result buffer only here and this is used a single
time, I decided to remove it also here.