building: Supported arbitrarily large files in onefile mode.
Previously, the maximum file size in CArchive was 2 GB, otherwise
the size value could not be stored. This was also not very
beneficial, because the bootloader allocates the size of the file
in memory during unpacking, which is a huge amount of data at once.
Now files larger than 512 MiB are divided into several archive entries
and thus unpacked to one file in several steps.