Update the assertion message in imphook.py
Allows the error from #4626 to be fixed by the user by making the problem clear.

This is a temporary fix, I'm going to submit another PR with a new cmdline option (and all that that entails) `--prefer-custom` so that when passed, PyInstaller will use custom hooks over PyInstaller ones, and when not passed, custom hooks that conflict with default ones will be ignored - not passed will be the default.

Update the assertion message in imphook.py, so that the user knows what the cause of the AssertionError is, allowing them to remove custom hooks.