Remove support for python from the windows store (#4702)
Docs: Remove support for Python from the Windows store

When using python from the windows store with PyInstaller, there are permissions errors that require a major mod to the UAC system in the underlying windows kernel to work around. 

It's caused by the damned `WindowsApps` directory, which, for security reasons, even running as *admin* can't even *read* the directory. This means that when PyInstaller (trys to) copy the Python dlls and librarys, it fails. You can work around this, but the method risks jailbreaking, which voids warranty. 

This can be "fixed" by saying that we don't support that build of Python. (Note that that is actually standard CPython).