makespec: Fix regression when trying to avoid hard-coded paths.
os.path.commonpath() ca not compare relative and absolute paths, thus a simple
call to ``pyinstaller somefile.py`` crashed. There was already a check to
assert one of the paths to be absolute. I now added a check whether the other
path is absolute, too. If not, there is no need to compare a single path.

Also add a test-case for this.

Introduced in 042d21cec66308f39ef530bf0e6dbd5de3fa442c.
Closes #4807.