Bootloader: Use snprintf() to simplify some string length checks.
Instead of copying strings and checking whether the string si terminated with
a Null bytes, use snprintf, which tells how much space the new string would
have taken and compare this to the available space.

Since snprintf() add some overhead, This code might be a bit less efficient.
But this sie the bootloader, and most of the functions are called once only.