Bootloader: Fix splitName().
This segfaulted if the string to split does not contain a filename part (e.g.
`aaaa:`), since strtok() then returns NULL for the second part. When working
on handling this case (and other cases), I discovered it is easier to no use
strtok() but just search the colon and split the string.