building: Add Splash Build Target
To enable the splash screen at compile-time this build target needs
to be added, which bundles all necessary files for SplashWriter and
checks if the installed Tcl/Tk version is suitable for the application.
It is also responsible for creating the script the tcl interpreter in
the bootloader will execute in order to create the splash screen.

Check tuple elements instead of tuple