Bootloader: Add support for a splash screen based on Tcl/Tk
Implementing a one-image splash screen with support of printing text
onto the image at runtime. This text can be updated via an tcp socket
based IPC system. The necessary data for the splash screen is inside
an archive item with the ARCHIVE_ITEM_SPLASH type. On runtime the
Tcl/Tk binaries loaded into the application an calls to the API are
made.