Bootloader: Add progress update to pyi_launch_extract_binaries
To update the progress of the extraction at runtime on the image,
API calls to pyi_splash are made. Since the messages are send
asynchronously to the splash screen the name of the current TOC being
extracted is posted into a queue for displaying (the name is inside a
malloc'ed buffer and therefore persistent).