Archive: Add SplashWriter for writing the resources
To write all data for the splash screen inside a file, which will be
included in the CArchive, the SplashWriter class is added. This writer
writes the struct of SPLASH_DATA_HEADER followed by the image, script
and requirements list to a specified file, which can be read by the 
bootloaders pyi_splash module.