building/hooks: Add pyi_splash python module
To let the users program communicate with the splash screen the
bootloader creates if splash resources are found in the archive,
this module connects to the tcp based IPC system. This provides a
common interface for the python program to communicate with the
bootloader.
This commit also includes tests to check if the pyi_module was
correctly included and executed.

Enhance fake-module

Enhance error in test_pyi_splash