setup: Fix long-description for PyPI.
Basically: Only use the Readme.rst as the long-description.
Also explicitly set the long-description-content-type.

PyPI now checks whether the uploaded `long-description` renders with the
selected long-description-content-type, and is quite strict on this: It does
not support Sphiny, but only plain reStructuredText.

While we could add more dummies for Sphinx-roles used in the change-log (which
we already had for e.g. `issue`), this failed with »Unknown directive type
"toctree"«. So I decided to no longer include the change-log into the
long-description, but instead add a section to the Readme (which is the base
of the long-description, pointing to the change-log.

I also enhanced the Relead descriptoion in the wiki to include running `twine
check` to catch errors early.