utils/hooks: Fix crash when running from Python build directory.
When running from a Python build directory (without installing), the prefix
for Makefile and config.h is none of the ones tested. So this routine did not
find a possible prefix, leading to an IndexError.

We must not add `dirname(filename)` to the list of prefixes to be tested,
since this would always win the "longest matching prefix" contest.