WIP Tests: Add test for multipackage importing tkinter.
TODO: Use generic script? Merge with other test-cases?

Multipackage tests failed by by chance when building for Python 3.8, since
tkinter gets imported (see below). So add a test for this.

The actuall reason, these tests failed is: If the depending package contains
data files in a sub-directory, these are transferred to the dependency as
expected, but the path in the archive holds the basename only. See
3e0f53726c3e6ccbe3ff4aaf4907426501e5a2d9 to rational about the basename.

In Python 3.8, multiprocessing.utils has a function `_cleanup_tests()` for
cleaning up multiprocessing resources when multiprocessing tests are
completed. This function imports the Python `test.support` package, which
imports tkinter.