hooks: Add hook for PIL.ImageFilter: Don't pull in numpy.
numpy is an optional component.

Since numpy is huge, let's see we can life without this import-link, to save
space and build time.

I assume that an application, which wants to use numpy in PIL.ImageFilter,
already imports numpy elsewhere, so it will be bundled and PIL.ImageFilter
will be able to import it.