hooks: PySide2.QtWebEngineWidgets: fix search for extra NSS libraries
The glob used to search for extra NSS libraries on linux turns out to
have no matches on Fedora 32. The `/lib64/nss` directory exists, but
contains no libraries, and so adding `/lib64/nss/*.so` to binaries
results in the following error: `Unable to find "/lib64/nss/*.so" when
adding binary and data files.`

Therefore, we now evaluate the glob to see if it has any matches
before adding it to `binaries`.