hooks: PySide2.QtWebEngineWidgets: collect QtQmlModels.framework
Added QtQmlModels to the list of libraries for which .framework directory
needs to be collected on macOS. Without it, QWebEngineView fails to load
a page and spams console with messages about QtQmlModels library not
being loaded.

The issue is currently masked by #5058, but once that is resolved,
test_PySide2_QWebEngine will likely start failing with timeout on macOS,
because failing to load the page will prevent the timer for exiting
the test program from being triggered.

This applies only to the PySide2 hook, because the same issue has
already been addressed in PyQt5 hook by d08a42c.