Hooks: sysconfig: Don't collect pyconfig.h and makefile.
As of Python 3.x, the config header and makefile are not longer used.  Trying to
include them causes build errors in pyenv/venv/equivalents.  Applying this fixes
5018, fixes 4775, and closes 1545 (a cleanup issue).

On Windows:
 The pyconfig.h data is hardcoded directly into sysconfig.py.
 And the makefile does not exist.
On Unix:
 A Python extension module contains all the details. This module is
 a hidden import since 3.6 but is already marked as such in the hook.

The functions to find and parse the files are still left over in sysconfig.py
but they are almost fully replaced by functions to get configuration parameters
directly. Should a user attempt to invoke the parser in a PyInstaller build, it
will now fail. They should instead use get_config_var() which will always work
(and is the recommended usage anyway).