modulegraph: Add node type ExtensionPackage.
Introduce a new node type ExtensionPackage, which signals that a package's
`__init__` is an extension module.

While the latest changes fixed modulegraph to not go into an endless recursion
if a package's `__init__` is an extension module, there still was an issue:
The returned node was of type ``Package``, leading to ``PYZ()`` trying to load
and compile the source (to get "missing" the code-object). Since the file is
an extension module, compiling failed, of course.