Tests: Add regression test for package's `__init__` being extension.
While the latest changes fixed modulegraph to not go into an endless recursion
if a package's `__init__` is an extension module, there still was an issue:
The returned node was of type ``Package``, leading to ``PYZ()`` trying to load
and compile the source. Since the file is an extension module, compiling
fails, of course.

This change adds a test-case for this.