loader: prevent local directory from shadowing system library
When loading a system dynamic library via ctypes.CDLL() in a
frozen progam, a local directory in sys._MEIPASS with a clashing
basename ends up preventing the library from being loaded.

To fix this, use an os.path.isdir() to decide whether to use
frozen or original library name.

Fixes #5178.