Hooks: distutils: Fix no suitable dest location found for pyconfig.h.
Locating `pyconfig.h` and the `makefile` gets into a mess when using
certain environment managers (pyenv-virtualenv) because PyInstaller,
whilst trying to find an appropriate `dest` path to put the files in,
gets confused by `sys.prefix` (or its varients) not being a parent dir of
the config and makefiles.

As of Python 3.6, direct parsing of `pyconfig.h` and `makefile` are
replaced by a Python module generated on building Python so that
these files are no longer required.