compat: use errors='replace' when decoding stdout from executed commands
Use decode(..., errors='replace') when decoding stdout from the
process executed in exec_command(), exec_command_stdout(), and
exec_command_all(). This will replace any invalid characters with
? instead of raising UnicodeDecodeError.

The problem with the strict decoding is that we actually have no
control over the encoding of the data in stdout. The process we
execute may directly or indirectly spawn another process with
merged stdout (and even redirected stderr), and the data from that
process may use arbitrary encoding that differs from the one that
the original process is using. In such situations, strict decoding
will lead to UnicodeDecodeError in places where this is difficult
to handle (e.g., collect_submodules() call in a hook). Therefore,
replacing offending characters with ? seems like a lesser evil.

An example of the above scenario is happening in #2258. The
collect_submodules() call in the hook for zmq is triggering an
import for zmq.backend.cffi, which tries to compile a C extension
module by invoking MSC cl.exe. The latter returns localized error
message encoded in the local codepage (different from utf-8 in
default configurations), resulting in UnicodeDecodeError in
the hook.