hookutils: robustify the collect_submodules() function
Prevent the extra output to stdout during module import from
showing up in the modules list collected by collect_submodules().
The package-walking process now encloses the output module names
with special prefix and suffix, which allows us to discard extra
lines.

This fixes the issue with parts of error messages ending up in
the hidden imports (and consequently in warnings about these
spurious modules not being found), as reported in #4813.