tests: add test for pkg_resources package content listing
The test script tests for behavior of resource_exists(), resource_isdir(),
and resource_listdir() functions from pkg_resources package (which
in turn call the methods with same name in the provider class).

The idea is to run the test script twice, once as unfrozen python
script and once as a frozen program. In both cases, the test package
is once present as a plain package directory and once as a zipped egg
(generated on-the-fly from the source directory).

This way, we test behavior of the original provider (DefaultProvider
or ZipProvider) and the provider used within the frozen application
(which we will need to implement to replace the currently used
NullProvider).