Hooks: Fix matplotlib hooks to support matplotlib 3.3.0. (#5006)
This alters where matplotlib's ``mpl-data`` folder is put and how it
is found at runtime. This is to support matplotlib>=3.3.0 without
breaking older versions. Fixes #5004.

Previously, the data folder was places in the app's root and a runtime
hook injected the environment variable ``MATPLOTLIBDATA`` to point to
said folder. This variable was deprecated in ``matplotlib 3.1.0`` and
now raises an error in ``matplotlib 3.3.0``.

With this commit, hook-matplotlib.py puts the ``mpl-data`` folder where matplotlib is now
(almost) hard-coded to expect it to be; which is in the parent folder of wherever
``matplotlib.__file__`` points to. With this move, we no longer need a runtime hook.

When I say `almost hard-coded` above, I mean it appears like matplotlib meant to make it
flexible but have mangled it as it now raises a ``NameError`` (see issue #5004) which
presumably isn't intentional.