OSX: Fix to reactivate running app on "reopen"
This is a follow-up to #5276 in order to address issue #4234. It allows
apps to re-activate if they are already running and the user
double-clicks their .app icon again. Previous to this double-clicking
the app icon would cause nothing to happen if the app was
already-running. This is because the app's loader process was being sent
an `rapp` event (reopen application) which it would promptly ignore.

On OSX the expectation is that the already-running app gets activated
and sent to the foreground when a user re-opens it via an icon on the
desktop.  This PR fixes that situation.

Summary of changes:

- We catch `rapp` events now too, in addition to `odoc` and `GURL`.
- The `rapp` events cannot be forwarded, sadly.  This is because they
  can only be addressed *TO* a bundle-id -- but the subordinate app on
  macOS that we launch with the `child_pid` has no real bundle id!
- So, instead, since we can't forward the exact `rapp` event, we do the
  next best thing: we call `SetFrontProcess()` which does the effective
  same thing -- tells the App's UI to go to foreground and activate
  itself.
- The test was updated to create a situation where the app loses focus,
  and then doing "open" on the .app again brings it back.
- The C code was refactored a bit as well.

I hope this PR's diff is not too large -- most of the code is the same
-- I just had to do a slight refactor to more easily support more event
types.