God rid of call to deprecated function
It turns out the reason why we never get an "activate" event is because
the bootloader process lacks a window.  So when we get an 'rapp' we need
to also sent an 'actv' event to the child process.

If we do that things work as expected.  Finalluy fixed!

Also in this commit: some refactoring.