bootloader: tests: fix building of tests on Windows (MSC)
On Windows, tests need to be linked against additional libraries.

We also need to avoid building tests for windowed bootloader
variant, because of incompatible entry point; tests use main(),
while executable in windowed mode needs WinMain(). And there's
probably not much point in having CLI tests without console,
anyway.

This allows building the bootloader (and the tests) with Visual
Studio when cmocka is present, and fixes #5314.