bootloader: on Windows, avoid using ntohl() from ws2_32 library
On Windows, use built-in byte-swap macros instead of ntohl(), which
is available from winsock2 library. Using the latter requires the
bootloader to be linked against the winsock2 library, which may have
other unintended consequences.

On all other platforms, continue using ntohl() for now, but hide it
behind the pyi_be32toh() macro.