Disable fail-fast for macOS GitHub actions CI
Prevent one failing workflow from automatically cancelling the
rest of the workflows. The fail-fast behavior makes it difficult
to assess the impact of a failing test; i.e., whether the failure
is specific to a python version, a random one, etc.

Because the workflows run in parallel, fail-fast has limited
benefit, and even less so in situations when the failing test is
located at the end of the test suite.