tests/ci: Improvements for macOS CI (#5352)
Prevent one failing workflow from automatically cancelling the
rest of the workflows. The fail-fast behavior makes it difficult
to assess the impact of a failing test; i.e., whether the failure
is specific to a python version, a random one, etc.

Because the workflows run in parallel, fail-fast has limited
benefit, and even less so in situations when the failing test is
located at the end of the test suite.


Increase the delays/wait times in test_osx_custom_protocol_handler
from 2 seconds to 5 seconds. The former value appears to be
insufficient for our CI and leads to occassional test failures.