building: avoid using .pyo module suffix in noarchive mode
The .pyo file suffix is gone since python 3.5 (PEP-488), and using
it when storing bytecode-optimized module files outside the PYZ
archive (i.e., noarchive mode) will lead to module-not-found errors.

The PEP-488 replaced the .pyo suffix with .opt-1.pyc and .opt-2.pyc,
but for bytecode-only module distribution, it seems that we need
to use the plain .pyc - same as for non-optimized case.