building: makespec: add new command-line switches for hookutils functions
Add new command-line options that map to the corresponding functions
from the hookutils:
* --collect-submodules: collect_submodules()
* --collect-data: collect_data_files()
* --collect-binaries: collect_dynamic_libs()
* --collect-all: collect_all()
* --copy-metadata: copy_metadata()

The hookutils calls are added to the generated .spec files. They
may be specified multiple times, and their results supplement
the entries that are manually passed via existing --add-data,
--add-binary, and --hidden-import switches.

This allows hookutils calls to be added from command-line, without
having to manually edit the .spec file.