hooks: update keyring hook (#5245)
* expand test_keyring

Instead instead of just importing the keyring module, try to get a
password from it. Doing so will raise a keyring.errors.NoKeyringError
if backends are not properly collected.

This is skipped on Linux, due to the SecretStorage backend requiring an active D-BUS session
and initialized keyring, and may need to unlock the keyring via
UI prompt. These conditions are difficult to satisfy in automated
test environments, so just skip the test on linux.

* update keyring hook

Turn the existing hook into a top-level hook, i.e., rename it
from  hook-keyring.backends.py to hook-keyring.py.

In addition to collecting the backends, also copy the package's
metadata, which is needed for backend discovery at run-time.

Fixes #4569.