building: Don't crash if destination directory was not removed.
If the destination directory is locked (e.g. if the directory is opened in the
file browser on Windows) it can not be deleted, thus makedir() failed over
this existing directory. We now silently ignore existing directories when
creating them, since the important part is that the files are removed (which
does not ignore errors).