hookutils: collect_submodules: limit the impact of a failed subpackage import
Fix the scenario where a single failed subpackage import during
a package inspection by collect_submodules() prevents the remainder
of the package from being inspected/collected.

Instead of re-raising the exception and aborting the package-walking,
we now print a message to stderr and continue, in an attempt to
salvage the situation based on the following rationale: if the
problematic sub-package is required by the program, then the unfrozen
version would have hit the same problem (and there's nothing we
can do about it). On the other hand, if the problematic sub-package
is not required by the program, we now properly collect all other
sub-packages and allow the program to work, as opposed to breaking
it by not collecting the remaining sub-packages at all.