hooks: PySide2/PyQt5: fix the availability check for older Qt5 versions (#5425)
Replace the truthyness check of {pyside2,pyqt5}_library_info.version
with explicit not-None check. The check is meant to ensure that the
library is importable, however, the version information is not available
in older Qt5 versions (QLibraryInfo::version() was added in Qt 5.8).

Therefore, an empty version array indicates that the library is available,
but is too old to have the version information exposed. The library
failing to import, however, should result in corresponding library_info
structure to have version set to None.

Fixes #5381.