Hooks: Fix macOS code signing for GTK. [skip-ci]
Fixes #5422. GdkPixbufLoader is by default stored in a
lib/gdk-pixbuf-2.0/2.10.0/loaders directory and the current GdkPixbuf
hooks package it in the same directory structure. When trying to code
sign a macOS app, Apple treats directories in Contents/MacOS/ with a
period in the name as a bundle and causes the code signing to fail. This
change moves this library to lib/gdk-pixbuf/loaders.