docs: Remove undefined cross-reference.
Remove the only occurrence of an ":option:" cross-reference. This reference
caused a warning, since we have no targets defined for options.

The actual fix would be to rewrite help2rst.py to generate Sphinx markup as
described a
<https://www.sphinx-doc.org/en/master/usage/restructuredtext/domains.html#directive-program>.
Anyhow, as of today, this ":option:" cross-reference was the only one. Thus as
of today is not worth this effort and I decided to simply remove the this
cross-reference.

Closes #6221 [skip ci]