bindepend: suppress missing library messages for macOS system libraries
Starting with macOS Big Sur, the system libraries are hidden. This
results in a slew of error messages about missing libraries when
a shared library is scanned for imports. However, all these error
messages are harmless, because we avoid collecting system libraries
on macOS, nor do we need to further analyze them for dependencies
(as they can depend only on other system libraries). Therefore,
suppress the missing-library error messages for system libraries.